+++ LinqBenchmarks/Enumerable/FatReferenceType/EnumerableFastReferenceTypeAny.cs
@@ -3,7 +3,7 @@ using NetFabric.Hyperlinq;
 using StructLinq;
 using System.Linq;
 
-namespace LinqBenchmarks.Enumerable.Int32
+namespace LinqBenchmarks.Enumerable.FatReferenceType
 {
     public class EnumerableFastReferenceTypeAny: EnumerableFatReferenceTypeBenchmarkBase
     {
